implement a consistent interface for canvas session object
